Set JAVA_HOME rather than source Conda env
When running the Docker image with a command from the CLI a non-login
shell is used and .bashrc is not sourced. Consequently, the Conda
environment is not correctly configured. This led to Java using the
system JAVA_HOME directory causing Nextflow to fail at startup.
Previously this was fixed by manually sourcing the appropriate Conda
file, which .bashrc would do automatically for login shells. Setting
JAVA_HOME is a cleaner solution.